+++ example/lib/main.dart
@@ -62,8 +62,7 @@ class _MyHomePageState extends State<MyHomePage> {
             icon: Icon(Icons.apps),
             title: Text('Home'),
             activeColor: Colors.red,
-            textAlign: TextAlign.center,
-            tooltipText: "Home Sweet Home",
+            textAlign: TextAlign.center
           ),
           BottomNavyBarItem(
             icon: Icon(Icons.people),